// apps/searchneu/app/graduate/common/types.ts
export enum NUPathEnum {
    ND = "Natural/Designed World",
    EI = "Creative Express/Innov",
    IC = "Interpreting Culture",
    FQ = "Formal/Quant Reasoning",
    SI = "Societies/Institutions",
    AD = "Analyzing/Using Data",
    DD = "Difference/Diversity",
    ER = "Ethical Reasoning",
    WF = "1st Yr Writing",
    WD = "Adv Writ Dscpl",
    WI = "Writing Intensive",
    EX = "Integration Experience",
    CE = "Capstone Experience",
  }
// ...
  export type INEUReq = INEUAndReq | INEUOrReq | INEUReqCourse;
  /**
   * A SearchNEU AND prerequisite object.
   *
   * @param type   The type of the SearchNEU prerequisite.
   * @param values The prerequisites that must be completed for this prereq. to be
   *   marked as done.
   */
  export interface INEUAndReq {
    type: "and";
    values: INEUReq[];
  }
  
  /**
   * A SearchNEU OR prerequisite object.
   *
   * @param type   The type of the SearchNEU prerequisite.
   * @param values The prerequisites of which one must be completed for this
   *   prerequisite to be marked as done.
   */
  export interface INEUOrReq {
    type: "or";
    values: INEUReq[];
  }
// ...
  export interface Section {
    type: "SECTION";
    title: string;
    requirements: Requirement2[];
    minRequirementCount: number;
    warnings?: string[];
  }
  
  /** Represents a degree requirement that allows a Section to be completed. */
  export type Requirement2 =
    | IXofManyCourse
    | IAndCourse2
    | IOrCourse2
    | ICourseRange2
    | IRequiredCourse
    | Section;
  /**
   * Represents a requirement where X number of credits need to be completed from
   * a list of courses.
   *
   * @param type          The type of requirement.
   * @param numCreditsMin The minimum number of credits needed to fulfill a given section.
   * @param courses       The list of requirements that the credits can be fulfilled from.
   */
  export interface IXofManyCourse {
    type: "XOM";
    numCreditsMin: number;
    courses: Requirement2[];
  }
  
  /**
   * Represents an 'AND' series of requirements.
   *
   * @param type    The type of requirement.
   * @param courses The list of requirements, all of which must be taken to
   *   satisfy this requirement.
   */
  export interface IAndCourse2 {
    type: "AND";
    courses: Requirement2[];
  }
  
  /**
   * Represents an 'OR' set of requirements.
   *
   * @param type    The type of requirement.
   * @param courses The list of requirements, one of which can be taken to satisfy
   *   this requirement.
   */
  export interface IOrCourse2 {
    type: "OR";
    courses: Requirement2[];
  }
// ...
  export interface ICourseRange2 {
    type: "RANGE";
    subject: string;
    idRangeStart: number;
    idRangeEnd: number;
    exceptions: IRequiredCourse[];
  }
  
  /**
   * A single required course.
   *
   * @param classId - The numeric ID of the course.
   * @param subject - The subject that the course is concerned with, such as CS
   *   (Computer Science).
   */
  export interface IRequiredCourse {
    type: "COURSE";
    classId: number;
    subject: string;
    description?: string;
  }
// ...
  export interface ScheduleCourse2<T> {
    name: string;
    classId: string;
    subject: string;
    prereqs?: INEUAndReq | INEUOrReq;
    coreqs?: INEUAndReq | INEUOrReq;
    nupaths?: NUPathEnum[];
    numCreditsMin: number;
    numCreditsMax: number;
    id: T;
    generic?: boolean;
  }